# BoardGame.View/OpenAPIs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "BoardGame API",
    "description": "Api for testing",
    "version": "v1"
  },
  "paths": {
    "/api/Game/start": {
      "post": {
        "tags": [
          "Game"
        ],
        "parameters": [
          {
            "name": "scenarioType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ScenarioType"
            }
          },
          {
            "name": "playersCount",
            "in": "query",
            "schema": {
              "maximum": 4,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateGameCommandResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateGameCommandResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateGameCommandResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Game/{idGame}/status": {
      "get": {
        "tags": [
          "Game"
        ],
        "parameters": [
          {
            "name": "idGame",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CheckGameQueryResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckGameQueryResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckGameQueryResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Game/{idGame}/map": {
      "get": {
        "tags": [
          "Game"
        ],
        "parameters": [
          {
            "name": "idGame",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetMapQueryResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMapQueryResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMapQueryResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Player": {
      "post": {
        "tags": [
          "Player"
        ],
        "parameters": [
          {
            "name": "idGame",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "row",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "col",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TurnCellCommandResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnCellCommandResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnCellCommandResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Cell": {
        "type": "object",
        "properties": {
          "idTil": {
            "type": "integer",
            "format": "int32"
          },
          "tileType": {
            "$ref": "#/components/schemas/TileType"
          },
          "isOpen": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CheckGameQueryResult": {
        "type": "object",
        "properties": {
          "gameExist": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CreateGameCommandResult": {
        "type": "object",
        "properties": {
          "idGame": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "GetMapQueryResult": {
        "type": "object",
        "properties": {
          "idGame": {
            "type": "string",
            "nullable": true
          },
          "map": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Cell"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ScenarioType": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "TileType": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "TurnCellCommandResult": {
        "type": "object",
        "properties": {
          "isTurn": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}